+++ tests/polars/functions/test_row_to_names_polars.py
@@ -19,7 +19,7 @@ def test_separator_type(df):
     Raise if separator is not a string
     """
     with pytest.raises(TypeError, match="separator should be.+"):
-        df.janitor.row_to_names([1, 2], separator=1)
+        df.row_to_names([1, 2], separator=1)
 
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
@@ -28,7 +28,7 @@ def test_row_numbers_type(df):
     Raise if row_numbers is not an int/list
     """
     with pytest.raises(TypeError, match="row_numbers should be.+"):
-        df.janitor.row_to_names({1, 2})
+        df.row_to_names({1, 2})
 
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
@@ -40,12 +40,12 @@ def test_row_numbers_list_type(df):
     with pytest.raises(
         TypeError, match="entry in the row_numbers argument should be.+"
     ):
-        df.janitor.row_to_names(["1", 2])
+        df.row_to_names(["1", 2])
 
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names(df):
-    df = df.janitor.row_to_names(2)
+    df = df.row_to_names(2)
     assert df.columns[0] == "3.2346125"
     assert df.columns[1] == "3"
     assert df.columns[2] == "lion"
@@ -55,7 +55,7 @@ def test_row_to_names(df):
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_single_list(df):
     "Test output if row_numbers is a list, and contains a single item."
-    df = df.janitor.row_to_names([2])
+    df = df.row_to_names([2])
     assert df.columns[0] == "3.2346125"
     assert df.columns[1] == "3"
     assert df.columns[2] == "lion"
@@ -65,7 +65,7 @@ def test_row_to_names_single_list(df):
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_list(df):
     "Test output if row_numbers is a list."
-    df = df.janitor.row_to_names([1, 2])
+    df = df.row_to_names([1, 2])
     assert df.columns[0] == "2.456234_3.2346125"
     assert df.columns[1] == "2_3"
     assert df.columns[2] == "leopard_lion"
@@ -74,7 +74,7 @@ def test_row_to_names_list(df):
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_delete_this_row(df):
-    df = df.janitor.row_to_names(2, remove_rows=True)
+    df = df.row_to_names(2, remove_rows=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 1.234_523_45
@@ -85,7 +85,7 @@ def test_row_to_names_delete_this_row(df):
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_list_delete_this_row(df):
-    df = df.janitor.row_to_names([2], remove_rows=True)
+    df = df.row_to_names([2], remove_rows=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 1.234_523_45
@@ -96,7 +96,7 @@ def test_row_to_names_list_delete_this_row(df):
 
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_delete_above(df):
-    df = df.janitor.row_to_names(2, remove_rows_above=True)
+    df = df.row_to_names(2, remove_rows_above=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 3.234_612_5
@@ -108,7 +108,7 @@ def test_row_to_names_delete_above(df):
 @pytest.mark.parametrize("df", [df, df.lazy()])
 def test_row_to_names_delete_above_list(df):
     "Test output if row_numbers is a list"
-    df = df.janitor.row_to_names([2, 3], remove_rows_above=True)
+    df = df.row_to_names([2, 3], remove_rows_above=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 3.234_612_5
@@ -123,9 +123,7 @@ def test_row_to_names_delete_above_delete_rows(df):
     Test output for remove_rows=True
     and remove_rows_above=True
     """
-    df = df.janitor.row_to_names(
-        [2, 3], remove_rows=True, remove_rows_above=True
-    )
+    df = df.row_to_names([2, 3], remove_rows=True, remove_rows_above=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 2.456234
@@ -140,7 +138,7 @@ def test_row_to_names_delete_above_delete_rows_scalar(df):
     Test output for remove_rows=True
     and remove_rows_above=True
     """
-    df = df.janitor.row_to_names(2, remove_rows=True, remove_rows_above=True)
+    df = df.row_to_names(2, remove_rows=True, remove_rows_above=True)
     if isinstance(df, pl.LazyFrame):
         df = df.collect()
     assert df.to_series(0)[0] == 1.23452345
@@ -157,4 +155,4 @@ def test_row_to_names_delete_above_list_non_consecutive(df):
     msg += "or the integers in a list are consecutive increasing, "
     msg += "with a difference of 1."
     with pytest.raises(ValueError, match=msg):
-        df.janitor.row_to_names([1, 3], remove_rows_above=True)
+        df.row_to_names([1, 3], remove_rows_above=True)