+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - [ENH] `xlsx_table` function now supports polars - Issue #1352 @samukweku
 - [ENH] Added a `clean_names` method for polars - it can be used to clean the column names, or clean column values . Issue #1343 @samukweku
 - [ENH] Improved performance for non-equi joins when using numba - @samukweku PR #1341
-
+- [ENH] pandas Index,Series, DataFrame now supported in the `complete` method. - PR #1369 @samukweku
 
 ## [v0.27.0] - 2024-03-21